GenAI_HW2/hw2_chatbot_v2/requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
python-docx
requests
beautifulsoup4
Pillow
mcp[cli]
